Done optimize scene + fix warning

+++ 04-Collision/Brick.cpp
@@ -5,8 +5,8 @@ Brick::Brick(int X, int Y, int W, int H)
 {
 	texture = new Load_img_file("Resources\\ground\\2.png", 1, 1, 1, 0);
 	sprite = new Load_resources(texture, 1000);
-	this->x = X;
-	this->y = Y;
+	this->x = (float)X;
+	this->y = (float)Y;
 	this->width = W;
 	this->height = H;
 	obj_type = def_ID::BRICK;
@@ -18,7 +18,7 @@ void Brick::Render(Camera * camera)
 
 	for (int i = 0; i < (int)ceil(width / BRICK_FRAME_WIDTH); i++)
 		for (int j = 0; j < (int)ceil(height / BRICK_FRAME_HEIGHT); j++)
-			sprite->Draw(pos.x + i * BRICK_FRAME_WIDTH, pos.y + j * BRICK_FRAME_HEIGHT);
+			sprite->Draw((int)pos.x + i * BRICK_FRAME_WIDTH, (int)pos.y + j * BRICK_FRAME_HEIGHT);
 
 	if (IS_DEBUG_RENDER_BBOX)
 		RenderBoundingBox(camera);
