+++ 04-Collision/FontGame.h
@@ -14,5 +14,5 @@ private:
 public:
 	FontGame();
 	~FontGame();
-	void Draw(float x, float y, string s);
+	void Draw(int x, int y, string s);
 };
