+++ 04-Collision/Items.cpp
@@ -22,7 +22,7 @@ void Items::Render(Camera * camera)
 
 	D3DXVECTOR2 pos = camera->Translate(x, y);
 
-	sprite->Draw(pos.x, pos.y);
+	sprite->Draw((int)pos.x, (int)pos.y);
 
 	if (IS_DEBUG_RENDER_BBOX)
 		RenderBoundingBox(camera);
@@ -77,7 +77,7 @@ void Whip::Update(DWORD dt, vector<LPGAMEOBJECT> * listObject)
 	vector<LPGAMEOBJECT> listObject_Brick;
 	listObject_Brick.clear();
 
-	for (int i = 0; i < listObject->size(); i++)
+	for (UINT i = 0; i < listObject->size(); i++)
 		if (listObject->at(i)->GetType() == def_ID::BRICK)
 			listObject_Brick.push_back(listObject->at(i));
